Refactor: 스타일 리팩토링
